rename crc32.h to avoid include conflicts with nxdk
